// models/information/listage/Information.ts
export default interface Information {
  id_information: string;
  num_matricule: string;
  photo_profil: string;
  nom: string;
  prenoms: string;
  sexe: string;
  date_naissance: string;
  lieu_naissance: string;
  nationalite: string;
  adresse: string;
  telephone: string;
  email: string;
  annee_universitaire: string;
  id_niveau: string;
  niveau: string;
  groupe: string;
  id_obs: string;
  admis: string;
  situation: string;
  date_insc: string;
  date_arret: string;
}